requirements/base.txt
Django>=2.0,<4.0

cx-Oracle==8.3.0
djangorestframework<4.0
djangorestframework-jsonapi==5.0.0
whitenoise==6.4.0
requests<=2.28.2
PyJWT==2.6.0
dataclasses-json==0.5.7
django-concurrency==2.4
python-dateutil==2.8.2
django-cors-headers==3.14.0
django-filter==22.1
tenacity==8.2.2
openpyxl==3.1.2
numpy==1.24.2
pandas==1.5.3
Unidecode==1.3.6
XlsxWriter==3.1.0
XlsxWriter==3.1.0
Pillow==9.5.0
xhtml2pdf==0.2.1
holidays==0.11.1
qrcode==7.4.2
